requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
asyncio-dgram==2.1.2
attrs==21.2.0
chardet==4.0.0
click==7.1.2
discord.py==1.7.3
dnspython==2.2.1
idna==3.3
mcstatus==9.0.0
multidict==5.2.0
python-dotenv==0.19.2
six==1.14.0
typing_extensions==4.0.1
yarl==1.7.2
